market/requirements.txt
asgiref==3.5.0
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
zipp==3.8.0